Updated Accorion view to use Accordion::Builder
