+++ lib/phlex_ui/accordion.rb
@@ -6,17 +6,6 @@ module PhlexUI
             div(**attrs, &)
         end
 
-        def add_item(title, content, open: false)
-            render Accordion::Item.new(open: open) do
-                render PhlexUI::Accordion::Trigger.new do
-                    render PhlexUI::Accordion::DefaultTrigger.new { title }
-                end
-                render PhlexUI::Accordion::Content.new do
-                    render PhlexUI::Accordion::DefaultContent.new { content }
-                end
-            end
-        end
-
         private
 
         def default_attrs
